SIGMA SGIMA SIGMA BOY

+++ src/main/java/hybrid/api/rendering/ScreenBounds.java
@@ -81,7 +81,7 @@ public class ScreenBounds {
         return "x " + x + " y " + y + " width: " + width + " height " + height;
     }
 
-    public boolean contains(int mouseX, int mouseY) {
+    public boolean contains(double mouseX, double mouseY) {
         return mouseX >= x
                 && mouseX <= x + width
                 && mouseY >= y
